login out

+++ src/store/request.js
@@ -7,7 +7,10 @@ const state = store.getState()
 axios.interceptors.request.use(function (config) {
     // ny times api add authorization error, so hide it
     // config.headers.Authorization = "Bear " + state.user.accessToken;
-    store.dispatch(commonActionCreators.setLoading());
+    if(config.url !== '/auth/login' && config.url !== '/auth/register') {
+      store.dispatch(commonActionCreators.setLoading());
+    }
+    
     return config;
   }, function (error) {
     return Promise.reject(error);
@@ -17,7 +20,11 @@ axios.interceptors.response.use(function (response) {
     store.dispatch(commonActionCreators.removeLoading());
     return response;
   }, function (error) {
-      store.dispatch(commonActionCreators.removeLoading());
-      store.dispatch(commonActionCreators.setError(error));
-    return Promise.reject(error);
+     
+      if(error.response.config.url !== '/auth/login' && error.response.config.url !== '/auth/register') {
+        store.dispatch(commonActionCreators.removeLoading());
+        store.dispatch(commonActionCreators.setError(error));
+      }
+      
+      return Promise.reject(error);
   });
